Restructure the project.
The proxy is now splitted into an Hermit library and a daemon. The
daemon manages all isles and starts an Unix socket server at
/tmp/hermit_daemon. The command just sends packets to the daemon and
awaits their return. The protocol consists of the enums Action and
ActionResult. The encoding happens with serde and bincode.

+++ tools/hermit_proxy/src/hermit/utils.rs
@@ -31,9 +31,9 @@ pub fn cpufreq() -> Result<u32> {
     Err(Error::MissingFrequency)
 }
 
-pub fn parse_mem(mem: &str) -> Result<usize> {
+pub fn parse_mem(mem: &str) -> Result<u64> {
     let (num, postfix): (String, String) = mem.chars().partition(|&x| x.is_numeric());
-    let num = num.parse::<usize>().map_err(|_| Error::ParseMemory)?;
+    let num = num.parse::<u64>().map_err(|_| Error::ParseMemory)?;
 
     let factor = match postfix.as_str() {
         "E" | "e" => 1 << 60,
